refactor(core): extract concurrency helpers to shared module
